backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
mysqlclient>=2.2,<3.0
python-decouple>=3.8,<4.0
requests>=2.31,<3.0
gunicorn>=21.2,<23.0
